Initial contentlayer optimizations
